Move Model Mapper Loader to prod
