+++ PokemonTests/LoadPokemonFromRemoteUseCaseTests.swift
@@ -6,43 +6,7 @@
 //
 
 import XCTest
-
-public protocol HTTPClient {
-    typealias Result = Swift.Result<(Data, HTTPURLResponse), Error>
-    
-    func get(from url: URL, completion: @escaping (Result) -> Void)
-}
-
-public final class RemotePokemonLoader {
-    private let url: URL
-    private let client: HTTPClient
-    
-    public enum Error: Swift.Error {
-        case unexpectedError
-        case invalidData
-    }
-    
-    public init(url: URL, client: HTTPClient) {
-        self.url = url
-        self.client = client
-    }
-    
-    private static var OK_200: Int { 200 }
-    
-    public func load(completion: @escaping (Result<String, Error>) -> Void) {
-        client.get(from: url) { result in
-            switch result {
-            case let .success((data, response)):
-                guard response.statusCode == Self.OK_200 else {
-                    return completion(.failure(.invalidData))
-                }
-                
-            case .failure:
-                completion(.failure(.unexpectedError))
-            }
-        }
-    }
-}
+import Pokemon
 
 final class LoadPokemonFromRemoteUseCaseTests: XCTestCase {
